move hasBeenRun and setup stuff to new setup script
